refactor(Archive): FilterSelector 컴포넌트 분리
