+++ src/pages/Archive/Archive.tsx
@@ -7,6 +7,7 @@ import Description from './components/Description';
 import PinnedQuestionList from './components/PinnedQuestionList';
 import QuestionList from './components/QuestionList';
 import useSectionScroll from './hooks/useSectionScroll';
+import FilterSelector from './components/FilterSelector';
 
 export interface AnswerItem {
   answerId: number;
